Remove the last two codec shifts, both on the decompression side
Neither is reached by the unit tests; both show up when funpack
decompresses a tile.  readint/readlonglong rebuild a value from
untrusted bytes and shift into the sign bit, and fits_rdecomp shifts
by the full width of the type when nbits is zero (the operand is
always zero there, so the value is unchanged).

+++ fits_hdecompress.c
@@ -2440,7 +2440,7 @@ unsigned char b[4];
 	 */
 	for (i=0; i<4; i++) qread(infile,(char *) &b[i],1);
 	a = b[0];
-	for (i=1; i<4; i++) a = (a<<8) + b[i];
+	for (i=1; i<4; i++) a = (int) ((((unsigned int) a) << 8) + b[i]);
 	return(a);
 }
 
@@ -2462,7 +2462,7 @@ unsigned char b[8];
 	 */
 	for (i=0; i<8; i++) qread(infile,(char *) &b[i],1);
 	a = b[0];
-	for (i=1; i<8; i++) a = (a<<8) + b[i];
+	for (i=1; i<8; i++) a = (LONGLONG) ((((ULONGLONG) a) << 8) + b[i]);
 	return(a);
 }
